requirements.txt
asgiref==3.8.1
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.3
django-countries==7.6.1
django-phonenumber-field==8.1.0
django-storages==1.14.6
idna==3.10
jmespath==1.0.1
phonenumbers==9.0.8
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
gunicorn==23.0.0
packaging==25.0
whitenoise==6.9.0
django-anymail[resend]
